refactor: components now contain the definition to provide to the core package
